FrontEnd/requirements.txt
pandas
psycopg2-binary
streamlit
plotly